[IMP] mail: clean up code of RTC join/leave
Part of task-2366708

closes odoo/odoo#76528

X-original-commit: 9a05dfe06d9d5706d0e45ad14d6ef8902d6b276e

+++ addons/mail/static/src/models/thread/thread.js
@@ -200,6 +200,12 @@ function factory(dependencies) {
             if ('id' in data) {
                 data2.id = data.id;
             }
+            if ('invitedGuests' in data) {
+                data2.invitedGuests = data.invitedGuests;
+            }
+            if ('invitedPartners' in data) {
+                data2.invitedPartners = data.invitedPartners;
+            }
             if ('is_minimized' in data && 'state' in data) {
                 data2.serverFoldState = data.is_minimized ? data.state : 'closed';
             }
@@ -261,8 +267,8 @@ function factory(dependencies) {
             if ('rtc_inviting_session' in data) {
                 data2.rtcInvitingSession = insert(data.rtc_inviting_session);
             }
-            if ('rtc_sessions' in data) {
-                data2.rtcSessions = insert(data.rtc_sessions.map(record => this.messaging.models['mail.rtc_session'].convertData(record)));
+            if ('rtcSessions' in data) {
+                data2.rtcSessions = data.rtcSessions;
             }
             if ('seen_partners_info' in data) {
                 if (!data.seen_partners_info) {
@@ -325,8 +331,6 @@ function factory(dependencies) {
                 this.messaging.soundEffects.channelLeave.play();
             }
             this.update({
-                invitedGuests: clear(),
-                invitedPartners: clear(),
                 mailRtc: unlink(),
                 rtcInvitingSession: unlink(),
             });
@@ -776,12 +780,15 @@ function factory(dependencies) {
                     channel_id: this.id,
                 },
             }, { shadow: true }));
+            if (!this.exists()) {
+                return;
+            }
             this.update({
                 mailRtc: link(this.messaging.mailRtc),
                 rtcInvitingSession: unlink(),
-                rtcSessions: insertAndReplace(rtcSessions.map(record => this.messaging.models['mail.rtc_session'].convertData(record))),
-                invitedGuests: insertAndReplace(invitedGuests),
-                invitedPartners: insertAndReplace(invitedPartners),
+                rtcSessions,
+                invitedGuests,
+                invitedPartners,
             });
             await this.async(() => this.messaging.mailRtc.initSession({
                 currentSessionId: sessionId,
@@ -806,9 +813,7 @@ function factory(dependencies) {
          */
         updateRtcSessions(rtcSessions) {
             const oldCount = this.rtcSessions.length;
-            this.update({
-                rtcSessions: insertAndReplace(rtcSessions.map(record => this.messaging.models['mail.rtc_session'].convertData(record)))
-            });
+            this.update({ rtcSessions });
             if (this.mailRtc) {
                 const newCount = this.rtcSessions.length;
                 if (newCount > oldCount) {
